Issue 20046  http://codereview.appspot.com/20046 Prevent labels that end in __ attached to loops and switches.
We don't rely on labels not being in the restricted namespace, but we
do enforce that in one place:
  foo__: 1
is illegal, but prior to this change
  foo__: while (1);
was not because the while loop is an instanceof of the abstract class
LabeledStatement, but not of the concrete class LabeledStmtWrapper.

Also, we did not check labels in break and continue.

R=erights@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3284 360991c0-9f3b-0410-ba87-8b362a00edf4