Issue 40105  http://codereview.appspot.com/40105 XSS in testbed output
The testbed has an XSS vulnerability when the embeddable flag is off.

As an example,
<IMG alt=foo&#x3c;/script&#x3e;&#x3c;script&#x3e;alert(1)&#x3c;/script&#x3e;>
causes an alert.

This affects only the testbed since other clients of the gadget
rewriter use the default behavior which is to render JS in embeddable
mode.
The testbed allows this to be overridden so that output is more
readable.

R=jasvir@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3446 360991c0-9f3b-0410-ba87-8b362a00edf4