uri normalization breaks several uris http://codereview.appspot.com/88118
When using a PluginEnvironment that allows URIs without rewriting,
the cajoler will turn this:
  <a href="mailto:a@b"></a>
  <a href="http://a.b/c;_d=e"></a>
into this:
  <a href="mailto:/a%40b"></a>
  <a href="http://a.b/c%3b%5fd%3de"></a>
which doesn't mean the same thing at all.

What's happening is the TemplateCompiler gets
a URI String from the PluginEnvironment,
then runs that String through normalizeUri,
which is a little buggy and mangles the URI 

I don't see any particular reason to mistrust the
string returned by the PluginEnvironment.
So rather than trying to fix normalizeUri,
I've removed the call to normalizeUri.

I've also fixed the error message for invalid uris.

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3583 360991c0-9f3b-0410-ba87-8b362a00edf4