domita_test fixes for IE http://codereview.appspot.com/95079
1. domita_test counting was slightly off.
this change fixes the invariant count==(pass+fail).

it's still more fragile than I like.  I have a change
in mind to make it less fragile, but the change will
be large and invasive, so not right now.

also, it's still missing accounting of the async assertions.

2. domita_test often crashes my ie6 and ie7.
it's not a reliable crash.  sometimes with no code
changes, it will mysteriously work.  this would
bother me more, but I've only seen it happen with
domita_test, not with anything else.

it seems to be something about the ScriptLoad test.
when I disable that test, I stop seeing IE crashes.
I'll open a bug to narrow this down more.

3. domita_test is sensitive to browser quirks in
formatting of rgb color values.
this change normalizes the color tests.

4. after appling this change and the
html-emitter change http://codereview.appspot.com/89076/show
domita_test has 3 fails on ie8, +1 on ie7, +2 on ie6.
I'll open a bug to describe those.

R=ihab.awad@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3600 360991c0-9f3b-0410-ba87-8b362a00edf4