fix some html-emitter problems on IE http://codereview.appspot.com/89076
This fixes
http://code.google.com/p/google-caja/issues/detail?id=1060
http://code.google.com/p/google-caja/issues/detail?id=1090

1. Cajoling this will fail to run on IE:

    <div id=a>a1</div>a2
    <script>1</script>
    <div id=b>b1</div>b2
    <script>2</script>
    c1

The error from IE is "'detached[...]' is null or not an object".

The reason is that html-emitter assumes that when you remove a
node from the document, node.parentNode will be null, but IE will
sometimes create another detached node to be the parent.  In this
example, one of the generated <span> elements has a
documentFragment parent.

This change fixes html-emitter so that when it searches
the detached nodes list, it makes sure it's comparing
the parentNode==null ancestor.

2. html-emitter gets confused if the source has
   <form id="x"><input name="id"></form>
because form.id is the <input>, not "x".

This change fixes that by using form.getAttribute('id')
instead of form.id.

R=mikesamuel@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3650 360991c0-9f3b-0410-ba87-8b362a00edf4