Oops, had some unused imports in CajitaRewriterTest.java. Fixed! http://codereview.appspot.com/115041
Change cajita.js prepareModule to be more paranoid about what it exposes
to the Cajita programmer. We wish to preserve the invariant that the module
object exposed to Cajita is transitively immutable.

In the old way, the debugging information on top-level modules was not being
frozen, and was being copied into the module exposed to Cajita, thus serving
as a communication channel.

R=maoziqing@gmail.com


git-svn-id: http://google-caja.googlecode.com/svn/trunk@3732 360991c0-9f3b-0410-ba87-8b362a00edf4